finishes preliminary live feed functionality

+++ tot-server-side/src/main/java/lt/verbus/totalisator/controller/dto/MatchDTO.java
@@ -25,9 +25,9 @@ public class MatchDTO {
 
     private Long totalisatorId;
 
-    private Byte homeScore;
+    private Integer homeScore;
 
-    private Byte awayScore;
+    private Integer awayScore;
 
     private String statusName;
